Day 14 / Star 2 - Clusterfuck
So this was a bit of a clusterfuck, spread over a couple of days.
Initial Star 2 code was fine... apart from incorrectly checking
if our string is contained, due to an 'optimization' for only checking
the end... And so we miss the first occurence of our input 360781 due
to the string actually ending with 3670810.

... So this set of a spiral of javascript memory / heap limit
exploration etc. Bumped into an issue of maximum array size of about 50m
numbers so implemented a layered array, which solved that issue.

Found the result much of course and got the wrong result. Figured out
the issue by running my input against another's internet solution to check for
validity of input + expected output by puzzle.

So in the end:
* There was no need to worry about hitting array size limits
* The data structure should have likely been a list
* Elves should have just been 1 and 2, not an array but that was just
speculation when doing Part 1